requirements.txt
discord.py
python-dotenv
python-dateutil
git+git://github.com/Klemek/miniscord.git
